cannon: Move hex utility
