+++ cannon/mipsevm/program/metadata.go
@@ -64,14 +64,3 @@ func (m *Metadata) SymbolMatcher(name string) func(addr uint32) bool {
 		return false
 	}
 }
-
-// HexU32 to lazy-format integer attributes for logging
-type HexU32 uint32
-
-func (v HexU32) String() string {
-	return fmt.Sprintf("%08x", uint32(v))
-}
-
-func (v HexU32) MarshalText() ([]byte, error) {
-	return []byte(v.String()), nil
-}
